src/load_data.py
from sqlalchemy import create_engine, text
from urllib.parse import quote_plus
import os
from pathlib import Path
import pandas as pd
from dotenv import load_dotenv

import logging

logging.basicConfig(
    level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
)

env_path = Path(__file__).parent.parent / "config" / ".env"
load_dotenv(dotenv_path=env_path)

user = os.getenv("DB_USER")
password = os.getenv("DB_PASSWORD")
host = os.getenv("DB_HOST")
port = os.getenv("DB_PORT")
database = os.getenv("DB_NAME")
def get_engine():
    logging
    return create_engine(
        f"postgresql+psycopg2://{user}:{quote_plus(password)}@{host}:{port}/{database}"
    )

engine = get_engine()

def load_weather_data(table_name:str, df):
    df.to_sql(name=table_name, con=engine, if_exists="append", index=False) 
    
    logging.info(f"Data loaded into {table_name} successfully.")
    
    df_check = pd.read_sql_query(text(f"SELECT * FROM {table_name}"), con=engine)
    logging.info(f"Data in {table_name}:\n{df_check.head()}")
    